requirements.txt
asgiref==3.4.1
Django==3.2.5
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pytz==2021.1
sqlparse==0.4.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
idna==3.2
multidict==5.1.0
pydantic==1.8.2
requests==2.26.0
tinvest==3.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3